+++ app/src/main/java/org/mozilla/fenix/components/Services.kt
@@ -21,7 +21,7 @@ import org.mozilla.fenix.components.metrics.Event
 import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.ext.getPreferenceKey
 import org.mozilla.fenix.settings.SupportUtils
-import org.mozilla.fenix.test.Mockable
+import org.mozilla.fenix.utils.Mockable
 
 /**
  * Component group which encapsulates foreground-friendly services.
